Refactor Model; add note input
I had to refactor the model because using a tuple was just getting out
of hand. I'm still not satisfied with the data structure of the model
but this is at least better. I think the solution is to improve the
architecture of the app to have a top-level model and a component
model, but since the goal is such a simple tool I don't think that's
the best route.

Works towards #5